+++ facturapi-net/FacturapiClient.cs
@@ -12,6 +12,8 @@ namespace Facturapi
         public Wrappers.ProductWrapper Product { get; private set; }
         public Wrappers.InvoiceWrapper Invoice { get; private set; }
         public Wrappers.OrganizationWrapper Organization { get; private set; }
+        public Wrappers.ReceiptWrapper Receipt { get; private set; }
+        public Wrappers.RetentionWrapper Retention { get; private set; }
 
         public FacturapiClient(string apiKey)
         {
@@ -19,12 +21,8 @@ namespace Facturapi
             this.Product = new Wrappers.ProductWrapper(apiKey);
             this.Invoice = new Wrappers.InvoiceWrapper(apiKey);
             this.Organization = new Wrappers.OrganizationWrapper(apiKey);
+            this.Receipt = new Wrappers.ReceiptWrapper(apiKey);
+            this.Retention = new Wrappers.RetentionWrapper(apiKey);
         }
-    }
-
-    [Obsolete("Deprecated. Use FacturapiClient instead.")]
-    public class Wrapper : FacturapiClient
-    {
-        public Wrapper(string apiKey) : base(apiKey) { }
     }
 }