+++ CHANGELOG.md
@@ -5,6 +5,30 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.1] - 2021-07-28
+
+### Added
+
+- Missing property on Stamp object: ComplementString
+
+## [2.0.0] - 2021-04-19
+
+### Breaking changes
+
+- Removed deprecated property `Invoice.ForeignTrade`.
+- Removed deprecated `Wrapper` class.
+
+### Added
+
+- Support for Receipts API
+- Support for Retentions API
+- Added missing properties on `Invoice` class: `ExternalId`, `Type`, `Stamp` and `Complements`.
+
+## [1.0.3] - 2020-08-19
+### Fixed
+
+- `Invoice.Payments[].Related` should be a List
+
 ## [1.0.2] - 2019-09-12
 ### Fixed